+++ programs/dynamic-bonding-curve/src/instructions/initialize_pool/process_initialize_virtual_pool_with_token2022.rs
@@ -2,8 +2,12 @@ use super::InitializePoolParameters;
 use crate::constants::MIN_LOCKED_LIQUIDITY_BPS;
 use crate::token::transfer_lamports_from_user;
 use crate::{
-    activation_handler::get_current_point, const_pda, state::TokenType,
-    token::update_account_lamports_to_minimum_balance, ConfigAccountLoader, PoolError,
+    activation_handler::get_current_point,
+    base_fee::BaseFeeEnumReader,
+    const_pda,
+    state::{BaseFeeMode, TokenType},
+    token::update_account_lamports_to_minimum_balance,
+    ConfigAccountLoader, PoolError,
 };
 use anchor_lang::prelude::*;
 use anchor_lang::solana_program::clock::SECONDS_PER_DAY;
@@ -42,9 +46,16 @@ pub fn process_initialize_virtual_pool_with_token2022<'info>(
         PoolError::InvalidMigrationLockedLiquidity
     );
 
+    require!(
+        config.pool_fees.base_fee.get_base_fee_mode()? != BaseFeeMode::RateLimiter,
+        PoolError::DeprecatedBaseFeeMode
+    );
+
     // validate min base fee
     config.pool_fees.base_fee.validate_min_base_fee()?;
 
+    // we do not add a deprecated migration option check here since the MigrationOption::MeteoraDamm config
+    // can only be created with TokenType::SplToken, so the token type check will already rejects it.
     let token_type_value =
         TokenType::try_from(config.token_type).map_err(|_| PoolError::InvalidTokenType)?;
     require!(