+++ programs/dynamic-bonding-curve/src/instructions/swap/process_swap.rs
@@ -15,7 +15,9 @@ use crate::{
     params::swap::TradeDirection,
     remaining_accounts::{parse_transfer_hook_accounts, AccountsType, TransferHookAccountsInfo},
     state::fee::FeeMode,
-    token::{transfer_token_from_pool_authority, transfer_token_from_user},
+    token::{
+        transfer_token_from_pool_authority, transfer_token_from_user, validate_transfer_fee_is_zero,
+    },
     ConfigAccountLoader, PoolAccountLoader, PoolError,
 };
 use anchor_lang::prelude::*;
@@ -282,6 +284,8 @@ pub fn process_swap<'a: 'info, 'info>(
         TradeDirection::BaseToQuote => (transfer_hook_base_accounts, None),
         TradeDirection::QuoteToBase => (None, transfer_hook_base_accounts),
     };
+
+    validate_transfer_fee_is_zero(&quote_mint.to_account_info())?;
 
     // send to reserve
     transfer_token_from_user(