+++ programs/dynamic-bonding-curve/src/utils/token.rs
@@ -4,9 +4,9 @@ use anchor_lang::{
     solana_program::program::{invoke, invoke_signed},
     solana_program::system_instruction::transfer,
 };
-use anchor_spl::associated_token::get_associated_token_address_with_program_id;
-use anchor_spl::token::accessor;
-use anchor_spl::token_2022::spl_token_2022::extension::transfer_hook;
+use anchor_spl::token_2022::spl_token_2022::extension::{
+    transfer_fee::TransferFeeConfig, transfer_hook,
+};
 use anchor_spl::{
     token::Token,
     token_2022::spl_token_2022::{
@@ -18,8 +18,8 @@ use anchor_spl::{
 use num_enum::{IntoPrimitive, TryFromPrimitive};
 
 use crate::const_pda::pool_authority::BUMP;
-use crate::safe_math::{SafeCast, SafeMath};
-use crate::state::{PoolState, PoolType};
+use crate::safe_math::SafeMath;
+use crate::state::{PoolState, TokenBadge};
 use crate::PoolError;
 
 #[derive(
@@ -40,22 +40,6 @@ pub fn get_token_program_flags<'a, 'info>(
         TokenProgramFlags::TokenProgram
     } else {
         TokenProgramFlags::TokenProgram2022
-    }
-}
-
-pub fn get_token_program_from_flag(token_program_flag: u8) -> Result<Pubkey> {
-    let token_program_flag: TokenProgramFlags = token_program_flag.safe_cast()?;
-    match token_program_flag {
-        TokenProgramFlags::TokenProgram => Ok(anchor_spl::token::ID),
-        TokenProgramFlags::TokenProgram2022 => Ok(anchor_spl::token_2022::ID),
-    }
-}
-
-pub fn get_token_program_from_pool_type(pool_type: u8) -> Result<Pubkey> {
-    let pool_type: PoolType = pool_type.safe_cast()?;
-    match pool_type {
-        PoolType::SplToken => Ok(anchor_spl::token::ID),
-        PoolType::Token2022 => Ok(anchor_spl::token_2022::ID),
     }
 }
 
@@ -185,18 +169,60 @@ pub fn transfer_token_from_pool_authority<'info>(
     Ok(())
 }
 
+fn is_transfer_fee_zero(mint: &StateWithExtensions<spl_token_2022::state::Mint>) -> Result<bool> {
+    if let Ok(transfer_fee_config) = mint.get_extension::<TransferFeeConfig>() {
+        let older_transfer_fee_bps = u16::from(
+            transfer_fee_config
+                .older_transfer_fee
+                .transfer_fee_basis_points,
+        );
+        let newer_transfer_fee_bps = u16::from(
+            transfer_fee_config
+                .newer_transfer_fee
+                .transfer_fee_basis_points,
+        );
+        return Ok(older_transfer_fee_bps == 0 && newer_transfer_fee_bps == 0);
+    }
+
+    Ok(true)
+}
+
+pub fn validate_transfer_fee_is_zero(mint_account_info: &AccountInfo) -> Result<()> {
+    if mint_account_info.owner.eq(&Token::id()) {
+        return Ok(());
+    }
+
+    let mint_data = mint_account_info.try_borrow_data()?;
+    let mint = StateWithExtensions::<spl_token_2022::state::Mint>::unpack(&mint_data)?;
+    require!(
+        is_transfer_fee_zero(&mint)?,
+        PoolError::QuoteMintHasNonZeroTransferFee
+    );
+
+    Ok(())
+}
+
+/// Rule: quote mint must be SPL-Token, or Token-2022 (non-native) with only metadata extensions
+/// never allow a non-zero transfer fee
 pub fn is_supported_quote_mint(mint_account: &InterfaceAccount<Mint>) -> Result<bool> {
     let mint_info = mint_account.to_account_info();
     if *mint_info.owner == Token::id() {
         return Ok(true);
     }
 
-    if spl_token_2022::native_mint::check_id(&mint_account.key()) {
-        return Err(PoolError::UnsupportNativeMintToken2022.into());
-    }
+    require!(
+        !spl_token_2022::native_mint::check_id(&mint_account.key()),
+        PoolError::UnsupportNativeMintToken2022
+    );
 
     let mint_data = mint_info.try_borrow_data()?;
     let mint = StateWithExtensions::<spl_token_2022::state::Mint>::unpack(&mint_data)?;
+
+    require!(
+        is_transfer_fee_zero(&mint)?,
+        PoolError::QuoteMintHasNonZeroTransferFee
+    );
+
     let extensions = mint.get_extension_types()?;
     for e in extensions {
         if e != ExtensionType::MetadataPointer && e != ExtensionType::TokenMetadata {
@@ -204,6 +230,29 @@ pub fn is_supported_quote_mint(mint_account: &InterfaceAccount<Mint>) -> Result<
         }
     }
     Ok(true)
+}
+
+pub fn validate_quote_mint_with_token_badge<'info>(
+    quote_mint: &InterfaceAccount<'info, Mint>,
+    token_badge: Option<&'info AccountInfo<'info>>,
+) -> Result<()> {
+    if !is_supported_quote_mint(quote_mint)? {
+        let token_badge = token_badge.ok_or_else(|| PoolError::InvalidTokenBadge)?;
+        require!(
+            is_token_badge_initialized(quote_mint.key(), token_badge)?,
+            PoolError::InvalidTokenBadge
+        );
+    }
+    Ok(())
+}
+
+fn is_token_badge_initialized<'info>(
+    mint: Pubkey,
+    token_badge: &'info AccountInfo<'info>,
+) -> Result<bool> {
+    let token_badge: AccountLoader<'_, TokenBadge> = AccountLoader::try_from(token_badge)?;
+    let token_badge = token_badge.load()?;
+    Ok(token_badge.token_mint == mint)
 }
 
 pub fn update_account_lamports_to_minimum_balance<'info>(
@@ -253,21 +302,5 @@ pub fn transfer_lamports_from_pool_account<'info>(
         PoolError::InsufficientPoolLamports
     );
 
-    Ok(())
-}
-
-pub fn validate_ata_token<'info>(
-    token_account: &AccountInfo<'info>,
-    owner: &Pubkey,
-    mint: &Pubkey,
-    token_program_id: &Pubkey,
-) -> Result<()> {
-    // validate ata address
-    let ata_address = get_associated_token_address_with_program_id(owner, mint, token_program_id);
-    require!(ata_address.eq(token_account.key), PoolError::IncorrectATA);
-
-    // validate owner
-    let current_owner = accessor::authority(token_account)?;
-    require!(current_owner.eq(owner), PoolError::IncorrectATA);
     Ok(())
 }