+++ tests/utils/index.ts
@@ -5,4 +5,6 @@ export * from "./types";
 export * from "./svm";
 export * from "./fetcher";
 export * from "./create_curve";
+export * from "./deprecatedRateLimiter";
+export * from "./deprecatedMeteoraDamm";
 export * from "./transferHookCounter";